requirements.txt
kaggle
pandas
google-cloud-storage
google-cloud-bigquery
db-dtypes
python-dotenv
pandas-gbq>=0.26.1